WEB_APP/ZSENSOR_COMUNICATION/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
dj-database-url==2.3.0
Django==4.2.16
django-heroku==0.3.1
djangorestframework==3.12.0
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
gunicorn==23.0.0
idna==3.10
packaging==24.2
pillow==10.4.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
PyJWT==2.9.0
pymongo==3.12.3
pyserial==3.5
python-decouple==3.8
pytz==2024.2
requests==2.32.3
sqlparse==0.5.2
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
whitenoise==6.8.2
